apps/reg-calculator/requirements-api.txt
# Additional requirements for API service
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
PyJWT>=2.8.0
pydantic>=2.0.0

